Behavior changes
Changed bubble column behavior for players:

If the player is on the ground and the bubble column is strong enough, then they will be pushed upwards off the ground unless they are in a dashing or crouching state.

If the player is crouching on the ground, then the strength of the wind is reset so that they are not immediately pushed with full force when they uncrouch.

Additionally fixed an issue with randomizing the bubble texture which has been present since the initial Spring Collab release.

Changes can be discussed and need further testing.